requirements.txt
tabulate
mysql-connector-python